Rename component beans
