+++ platform/src/main/java/org/stellar/anchor/platform/component/sep/ConfigBeans.java
@@ -2,11 +2,13 @@ package org.stellar.anchor.platform.component.sep;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.stellar.anchor.config.*;
+import org.stellar.anchor.platform.config.*;
 import org.stellar.anchor.platform.configurator.ConfigManager;
 import org.stellar.anchor.platform.configurator.SepConfigManager;
 
 @Configuration
-public class ConfigManagerBean {
+public class ConfigBeans {
   @Bean(name = "configManager")
   ConfigManager configManager() {
     return SepConfigManager.getInstance();
