+++ platform/src/main/java/org/stellar/anchor/platform/component/share/DataBeans.java
@@ -4,6 +4,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.stellar.anchor.platform.data.*;
 import org.stellar.anchor.platform.observer.stellar.JdbcStellarPaymentStreamerCursorStore;
+import org.stellar.anchor.platform.observer.stellar.PaymentObservingAccountStore;
 import org.stellar.anchor.sep24.Sep24TransactionStore;
 import org.stellar.anchor.sep38.Sep38QuoteStore;
 
@@ -28,5 +29,10 @@ public class DataBeans {
   JdbcStellarPaymentStreamerCursorStore stellarPaymentStreamerCursorStore(
       PaymentStreamerCursorRepo paymentStreamerCursorRepo) {
     return new JdbcStellarPaymentStreamerCursorStore(paymentStreamerCursorRepo);
+  }
+
+  @Bean
+  public PaymentObservingAccountStore observingAccountStore(PaymentObservingAccountRepo repo) {
+    return new PaymentObservingAccountStore(repo);
   }
 }
