+++ platform/src/main/java/org/stellar/anchor/platform/component/share/MetricsBeans.java
@@ -1,13 +1,25 @@
 package org.stellar.anchor.platform.component.share;
 
+import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.stellar.anchor.config.MetricConfig;
+import org.stellar.anchor.platform.config.PropertyMetricConfig;
 import org.stellar.anchor.platform.data.JdbcSep31TransactionRepo;
 import org.stellar.anchor.platform.service.MetricEmitterService;
 
 @Configuration
 public class MetricsBeans {
+
+  /**********************************
+   * Metrics configurations
+   */
+  @Bean
+  @ConfigurationProperties(prefix = "metrics")
+  MetricConfig metricConfig() {
+    return new PropertyMetricConfig();
+  }
+
   @Bean
   public MetricEmitterService metricService(
       MetricConfig metricConfig, JdbcSep31TransactionRepo sep31TransactionRepo) {
